feat(df_mlx): add enhance.py inference pipeline
- Add comprehensive enhance.py module for MLX DeepFilterNet4:
  - load_audio/save_audio with soundfile backend
  - resample utility with scipy
  - load_model for checkpoint loading
  - enhance() for single audio enhancement
  - enhance_streaming() for real-time processing
  - enhance_file() and enhance_batch() for file I/O
  - Full CLI interface with argparse

- Fix evaluation.py np.mean() return type (floating[Any] -> float)

- Update ROADMAP.md:
  - Mark Phase 1 (Training Foundations) complete
  - Mark Phase 2 (GAN Training) complete
  - Mark Phase 4 (Inference & Enhancement) complete
  - Update Phase 3 evaluation task status

Tested: load_audio, enhance, save_audio, streaming, CLI --help

+++ DeepFilterNet/df_mlx/evaluation.py
@@ -287,7 +287,7 @@ class Metric:
         """Compute mean of all results."""
         if not self.results:
             return 0.0
-        return np.mean([r.enhanced for r in self.results])
+        return float(np.mean([r.enhanced for r in self.results]))
 
     def clear(self):
         """Clear results."""
